refactor: 견적 요청 관련 컴포넌트 폴더 위치 변경
